Vitis-1444 API for aborting hanging command (#5592)
Add xrt::run::abort() to abort a run object.
Add example demonstrating use of abort() API.

Use uint64_t for exec_bo_handle. The handle is an 'fd' on Linux but a
void* handle on windows.

+++ src/runtime_src/core/include/ert.h
@@ -361,7 +361,7 @@ struct ert_abort_cmd {
   };
 
   /* payload */
-  uint32_t exec_bo_handle;
+  uint64_t exec_bo_handle;
 };
 
 /**
